D: 完善数学库，编译通过

+++ Akasha/engine/include/TypeDefines.h
@@ -2,8 +2,8 @@
 
 using uint64		= unsigned __int64;
 using int64			= __int64;
-using uint32		= unsigned int;
-using int32			= int;
+using uint32		= unsigned long int;
+using int32			= long int;
 using uint16		= unsigned short;
 using int16			= short;
 using uint8			= unsigned char;
